Fixed some issues in recent commits with project ids (#16963)
- session timeout kill hours is now a floating point for
better testing

Reviewed in: https://github.com/rstudio/rstudio-pro/pull/9979

+++ src/cpp/core/include/core/r_util/RActiveSessions.hpp
@@ -790,8 +790,8 @@ public:
    std::string userHomeDir() { return settings_.get("userHomeDir", ""); }
    void setUserHomeDir(const std::string& userHomeDir) { settings_.set("userHomeDir", userHomeDir); }
 
-   int sessionTimeoutKillHours() { return settings_.getInt("sessionTimeoutKillHours", 0); }
-   void setSessionTimeoutKillHours(int val) { settings_.set("sessionTimeoutKillHours", val); }
+   double sessionTimeoutKillHours() { return settings_.getDouble("sessionTimeoutKillHours", 0); }
+   void setSessionTimeoutKillHours(double val) { settings_.set("sessionTimeoutKillHours", val); }
 
    core::Error destroy() { return filePath_.removeIfExists(); }
